+++ src/server/commands/adb/manager.ts
