+++ src/server/commands/games/images.ts
