+++ src/server/commands/index.ts
@@ -1,4 +1,4 @@
-import { Command, CommandEvent } from "@comands/types";
+import { Command, CommandEvent } from "@commands/types";
 import { sendInfo } from "@server";
 import log from "@server/log";
