+++ src/server/commands/settings/index.ts
@@ -1,4 +1,4 @@
-import { Command, CommandEvent } from "@comands/types";
+import { Command, CommandEvent } from "@commands/types";
 import settingsManager from "./manager";
 
 export type SettingsCommandPayload = SettinsActions;
